+++ Magic/src/main/java/com/elmakers/mine/bukkit/wand/Wand.java
@@ -20,7 +20,6 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 import org.apache.commons.lang.StringUtils;
-import org.apache.commons.lang.WordUtils;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Color;
@@ -58,7 +57,6 @@ import com.elmakers.mine.bukkit.api.item.ItemData;
 import com.elmakers.mine.bukkit.api.magic.MageClassTemplate;
 import com.elmakers.mine.bukkit.api.magic.MageContext;
 import com.elmakers.mine.bukkit.api.magic.MageController;
-import com.elmakers.mine.bukkit.api.magic.MagicAttribute;
 import com.elmakers.mine.bukkit.api.magic.MagicProperties;
 import com.elmakers.mine.bukkit.api.magic.MagicPropertyType;
 import com.elmakers.mine.bukkit.api.magic.MaterialSet;
@@ -152,8 +150,6 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
     private String path = "";
     private String inventoryOpenLore = "";
     private List<String> mageClassKeys = null;
-    private boolean superProtected = false;
-    private boolean superPowered = false;
     private boolean glow = false;
     private boolean spellGlow = false;
     private boolean bound = false;
@@ -200,19 +196,6 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
     private MaterialAndData inactiveIcon = null;
     private int inactiveIconDelay = 0;
     private String upgradeTemplate = null;
-
-    protected float consumeReduction = 0;
-    protected float cooldownReduction = 0;
-    protected float costReduction = 0;
-    protected Map<String, Double> protection;
-    private float power = 0;
-    private float earnMultiplier = 1;
-
-    private float blockFOV = 0;
-    private float blockChance = 0;
-    private float blockReflectChance = 0;
-    private int blockMageCooldown = 0;
-    private int blockCooldown = 0;
 
     private int maxEnchantCount = 0;
     private int enchantCount = 0;
@@ -854,6 +837,7 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
 
     @Override
     public float getCostReduction() {
+        float costReduction = getFloat("cost_reduction");
         if (mage != null) {
             float reduction = mage.getCostReduction();
             return worn ? reduction : stackPassiveProperty(reduction, costReduction * controller.getMaxCostReduction());
@@ -863,6 +847,7 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
 
     @Override
     public float getCooldownReduction() {
+        float cooldownReduction = getFloat("cooldown_reduction");
         if (mage != null) {
             float reduction = mage.getCooldownReduction();
             return worn ? reduction : stackPassiveProperty(reduction, cooldownReduction * controller.getMaxCooldownReduction());
@@ -872,6 +857,7 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
 
     @Override
     public float getConsumeReduction() {
+        float consumeReduction = getFloat("consume_reduction");
         if (mage != null) {
             float reduction = mage.getConsumeReduction();
             return worn ? reduction : stackPassiveProperty(reduction, consumeReduction);
@@ -891,27 +877,27 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
 
     @Override
     public float getPower() {
-        return power;
+        return getFloat("power");
     }
 
     @Override
     public boolean isSuperProtected() {
-        return superProtected;
+        return getBoolean("protected");
     }
 
     @Override
     public boolean isSuperPowered() {
-        return superPowered;
+        return getBoolean("powered");
     }
 
     @Override
     public boolean isConsumeFree() {
-        return consumeReduction >= 1;
+        return getConsumeReduction() >= 1;
     }
 
     @Override
     public boolean isCooldownFree() {
-        return cooldownReduction > 1;
+        return getCooldownReduction() > 1;
     }
 
     @Override
@@ -2012,24 +1998,11 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
     protected void loadParameters() {
         // These should only be parameters to are OK and safe (performance-wise in particular)
         // to change on the fly without fully reloading the wand
-        consumeReduction = getFloat("consume_reduction");
-        cooldownReduction = getFloat("cooldown_reduction");
-        costReduction = getFloat("cost_reduction");
-        power = getFloat("power");
-        blockChance = getFloat("block_chance");
-        blockReflectChance = getFloat("block_reflect_chance");
-        blockFOV = getFloat("block_fov");
-        blockMageCooldown = getInt("block_mage_cooldown");
-        blockCooldown = getInt("block_cooldown");
-
-        earnMultiplier = getFloat("earn_multiplier", getFloat("sp_multiplier", 1));
         quietLevel = getInt("quiet");
         if (quietLevel == 0 && getBoolean("quiet")) {
             quietLevel = 1;
         }
         effectBubbles = getBoolean("effect_bubbles");
-        superPowered = getBoolean("powered");
-        superProtected = getBoolean("protected");
         glow = getBoolean("glow");
         spellGlow = getBoolean("spell_glow");
         undroppable = getBoolean("undroppable");
@@ -2267,13 +2240,6 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
             migrateProtection("protection_fire", "fire");
             migrateProtection("protection_explosions", "explosion");
             protectionConfig = getConfigurationSection("protection");
-        }
-
-        if (protectionConfig != null) {
-            protection = new HashMap<>();
-            for (String protectionKey : protectionConfig.getKeys(false)) {
-                protection.put(protectionKey, protectionConfig.getDouble(protectionKey));
-            }
         }
 
         hasId = getBoolean("unique", false);
@@ -2879,68 +2845,6 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
         return "<div style=\"background-color: black; margin: 8px; padding: 8px\">" + StringUtils.join(lore, "<br/>") + "</div>";
     }
 
-    private String getPropertyString(String templateName, float value) {
-        return getPropertyString(templateName, value, 1, false);
-    }
-
-    private String getPropertyString(String templateName, float value, float max, boolean defaultStack) {
-        String propertyTemplate = getBoolean("stack", defaultStack) ? "property_stack" : "property_value";
-        if (value < 0) {
-            propertyTemplate = propertyTemplate + "_negative";
-        }
-        return controller.getMessages().getPropertyString(getMessageKey(templateName), value, max, getMessageKey(propertyTemplate));
-    }
-
-    private String formatPropertyString(String message, float value, float max) {
-        String propertyTemplate = getBoolean("stack") ? "property_stack" : "property_value";
-        if (value < 0) {
-            propertyTemplate = propertyTemplate + "_negative";
-        }
-        return controller.getMessages().formatPropertyString(message, value, max, getMessage(propertyTemplate));
-    }
-
-    private void addDamageTypeLore(String property, String propertyType, double amount, List<String> lore) {
-        addDamageTypeLore(property, propertyType, amount, 1, lore);
-    }
-
-    private void addDamageTypeLore(String property, String propertyType, double amount, double max, List<String> lore) {
-        addDamageTypeLore(property, propertyType, amount, max, lore, null);
-    }
-
-    private void addDamageTypeLore(String property, String propertyType, double amount, double max, List<String> lore, String unknownDefault) {
-        if (amount != 0) {
-            String prefix = getMessageKey("prefixes." + property);
-            prefix = controller.getMessages().get(prefix, "");
-            String templateKey = getMessageKey(property + "." + propertyType);
-            String template;
-            if (controller.getMessages().containsKey(templateKey)) {
-                template = controller.getMessages().get(templateKey);
-            } else {
-                templateKey = getMessageKey(property + ".unknown");
-                template = controller.getMessages().get(templateKey);
-                String pretty = propertyType.substring(0, 1).toUpperCase() + propertyType.substring(1);
-                template = template.replace("$type", pretty);
-                if (unknownDefault != null && !unknownDefault.isEmpty()) {
-                    // This is some special-case hackery, currently only used for enchantments
-                    unknownDefault = WordUtils.capitalize(unknownDefault.toLowerCase().replace("_", " "));
-                    template = template.replace("$name", unknownDefault);
-                }
-            }
-            template = formatPropertyString(prefix + template, (float)amount, (float)max);
-            ConfigurationUtils.addIfNotEmpty(template, lore);
-        }
-    }
-
-    public String getLevelString(String templateName, float amount)
-    {
-        return controller.getMessages().getLevelString(getMessageKey(templateName), amount);
-    }
-
-    public String getLevelString(String templateName, float amount, float max)
-    {
-        return controller.getMessages().getLevelString(getMessageKey(templateName), amount, max);
-    }
-
     protected List<String> getCustomLore(Collection<String> loreTemplate) {
         List<String> lore = new ArrayList<>();
         for (String line : loreTemplate) {
@@ -3346,86 +3250,18 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
         addPropertyLore(lore, false);
     }
 
-    protected void addPropertyLore(List<String> lore, boolean isSingleSpell)
-    {
-        if (usesMana() && effectiveManaMax > 0) {
-            int manaMax = getManaMax();
-            if (effectiveManaMax != manaMax) {
-                String fullMessage = getLevelString("mana_amount_boosted", manaMax, controller.getMaxMana());
-                ConfigurationUtils.addIfNotEmpty(fullMessage.replace("$mana", Integer.toString((int)Math.ceil(effectiveManaMax))), lore);
-            } else {
-                ConfigurationUtils.addIfNotEmpty(getLevelString("mana_amount", manaMax, controller.getMaxMana()), lore);
-            }
-            int manaRegeneration = getManaRegeneration();
-            if (manaRegeneration > 0 && effectiveManaRegeneration > 0) {
-                if (effectiveManaRegeneration != manaRegeneration) {
-                    String fullMessage = getLevelString("mana_regeneration_boosted", manaRegeneration, controller.getMaxManaRegeneration());
-                    ConfigurationUtils.addIfNotEmpty(fullMessage.replace("$mana", Integer.toString((int)Math.ceil(effectiveManaRegeneration))), lore);
-                } else {
-                    ConfigurationUtils.addIfNotEmpty(getLevelString("mana_regeneration", manaRegeneration, controller.getMaxManaRegeneration()), lore);
-                }
-            }
-            float manaPerDamage = getFloat("mana_per_damage");
-            if (manaPerDamage > 0) {
-                ConfigurationUtils.addIfNotEmpty(getLevelString("mana_per_damage", manaPerDamage, controller.getMaxManaRegeneration()), lore);
-            }
-        }
-        if (blockReflectChance > 0) {
-            ConfigurationUtils.addIfNotEmpty(getLevelString("reflect_chance", blockReflectChance), lore);
-        } else if (blockChance != 0) {
-            ConfigurationUtils.addIfNotEmpty(getLevelString("block_chance", blockChance), lore);
-        }
-        float manaMaxBoost = getManaMaxBoost();
-        if (manaMaxBoost != 0) {
-            ConfigurationUtils.addIfNotEmpty(getPropertyString("mana_boost", manaMaxBoost, 1, true), lore);
-        }
-        float manaRegenerationBoost = getManaRegenerationBoost();
-        if (manaRegenerationBoost != 0) {
-            ConfigurationUtils.addIfNotEmpty(getPropertyString("mana_regeneration_boost", manaRegenerationBoost, 1, true), lore);
-        }
-
-        if (earnMultiplier > 1) {
-            String earnDescription = getPropertyString("earn_multiplier", earnMultiplier - 1);
-            String earnType = getController().getMessages().get("currency.sp.name_short", "SP");
-            earnDescription = earnDescription.replace("$type", earnType);
-            ConfigurationUtils.addIfNotEmpty(earnDescription, lore);
-        }
-
-        if (castSpell != null) {
-            SpellTemplate spell = controller.getSpellTemplate(castSpell);
-            if (spell != null)
-            {
-                ConfigurationUtils.addIfNotEmpty(getMessage("spell_aura").replace("$spell", spell.getName()), lore);
-            }
-        }
-        for (Map.Entry<PotionEffectType, Integer> effect : getPotionEffects().entrySet()) {
-            ConfigurationUtils.addIfNotEmpty(describePotionEffect(effect.getKey(), effect.getValue()), lore);
-        }
-
-        if (getBoolean("ignored_by_mobs")) {
-            ConfigurationUtils.addIfNotEmpty(getMessage("ignored_by_mobs"), lore);
-        }
-
+    protected void addPropertyLore(List<String> lore, boolean isSingleSpell) {
         // If this is a passive wand, then reduction properties stack onto the mage when worn.
         // In this case we should show it as such in the lore.
         if (worn) isSingleSpell = false;
 
-        if (consumeReduction != 0 && !isSingleSpell) ConfigurationUtils.addIfNotEmpty(getPropertyString("consume_reduction", consumeReduction), lore);
+        super.addPropertyLore(lore, isSingleSpell);
 
-        if (costReduction != 0 && !isSingleSpell) ConfigurationUtils.addIfNotEmpty(getPropertyString("cost_reduction", costReduction), lore);
-        if (cooldownReduction != 0 && !isSingleSpell) ConfigurationUtils.addIfNotEmpty(getPropertyString("cooldown_reduction", cooldownReduction), lore);
-        if (power > 0) ConfigurationUtils.addIfNotEmpty(getLevelString("power", power), lore);
-        if (superProtected) {
-            ConfigurationUtils.addIfNotEmpty(getMessage("super_protected"), lore);
-        } else if (protection != null) {
-            for (Map.Entry<String, Double> entry : protection.entrySet()) {
-                String protectionType = entry.getKey();
-                double amount = entry.getValue();
-                addDamageTypeLore("protection", protectionType, amount, lore);
+        if (castSpell != null) {
+            SpellTemplate spell = controller.getSpellTemplate(castSpell);
+            if (spell != null) {
+                ConfigurationUtils.addIfNotEmpty(getMessage("spell_aura").replace("$spell", spell.getName()), lore);
             }
-        }
-        if (superPowered) {
-            ConfigurationUtils.addIfNotEmpty(getMessage("super_powered"), lore);
         }
 
         if (isEnchantable()) {
@@ -3438,49 +3274,6 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
                     String[] pieces = StringUtils.split(enchantmentKey, ":");
                     enchantmentKey = pieces[pieces.length - 1];
                     addDamageTypeLore("enchantment", enchantmentKey, level, 0, lore, enchantmentKey);
-                }
-            }
-        }
-
-        ConfigurationSection weaknessConfig = getConfigurationSection("weakness");
-        if (weaknessConfig != null) {
-            Set<String> keys = weaknessConfig.getKeys(false);
-            for (String key : keys) {
-                addDamageTypeLore("weakness", key, weaknessConfig.getDouble(key), lore);
-            }
-        }
-
-        ConfigurationSection strengthConfig = getConfigurationSection("strength");
-        if (strengthConfig != null) {
-            Set<String> keys = strengthConfig.getKeys(false);
-            for (String key : keys) {
-                addDamageTypeLore("strength", key, strengthConfig.getDouble(key), lore);
-            }
-        }
-        ConfigurationSection attributes = getConfigurationSection("attributes");
-        if (attributes != null) {
-            // Don't bother with the lore at all if the template has been blanked out
-            String template = getMessage("attributes");
-            if (!template.isEmpty()) {
-                Set<String> keys = attributes.getKeys(false);
-                for (String key : keys) {
-                    String label = controller.getMessages().get("attributes." + key + ".name", key);
-
-                    // We are only display attributes as integers for now
-                    int value = attributes.getInt(key);
-                    if (value == 0) continue;
-
-                    float max = 1;
-                    MagicAttribute attribute = controller.getAttribute(key);
-                    if (attribute != null) {
-                        Double maxValue = attribute.getMax();
-                        if (maxValue != null) {
-                            max = (float)(double)maxValue;
-                        }
-                    }
-
-                    label = getPropertyString("attributes", value, max, true).replace("$attribute", label);
-                    lore.add(label);
                 }
             }
         }
@@ -4952,6 +4745,7 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
     }
 
     public boolean usesSP() {
+        float earnMultiplier = getFloat("earn_multiplier", getFloat("sp_multiplier", 1));
         return controller.isSPEarnEnabled() && hasSpellProgression && earnMultiplier > 0;
     }
 
@@ -5542,11 +5336,6 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
             }
             updateHotbarStatus();
             checkBossBar();
-        }
-
-        // Check for blocking cooldown
-        if (player.isBlocking() && blockMageCooldown > 0) {
-            mage.setRemainingCooldown(blockMageCooldown);
         }
 
         if (!worn) {
@@ -6978,31 +6767,13 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
     @Override
     public boolean isBlocked(double angle) {
         if (mage == null) return false;
-        if (blockChance == 0) return false;
-        if (blockFOV > 0 && angle > blockFOV) return false;
-        long lastBlock = mage.getLastBlockTime();
-        if (blockCooldown > 0 && lastBlock > 0 && lastBlock + blockCooldown > System.currentTimeMillis()) return false;
-        boolean isBlocked = Math.random() <= blockChance;
-        if (isBlocked) {
-            playEffects("spell_blocked");
-            mage.setLastBlockTime(System.currentTimeMillis());
-        }
-        return isBlocked;
+        return mage.isBlocked(angle);
     }
 
     @Override
     public boolean isReflected(double angle) {
         if (mage == null) return false;
-        if (blockReflectChance == 0) return false;
-        if (blockFOV > 0 && angle > blockFOV) return false;
-        long lastBlock = mage.getLastBlockTime();
-        if (blockCooldown > 0 && lastBlock > 0 && lastBlock + blockCooldown > System.currentTimeMillis()) return false;
-        boolean isReflected = Math.random() <= blockReflectChance;
-        if (isReflected) {
-            playEffects("spell_reflected");
-            if (mage != null) mage.setLastBlockTime(System.currentTimeMillis());
-        }
-        return isReflected;
+        return mage.isReflected(angle);
     }
 
     @Nullable